+++ src/main/java/com/zagurskaya/cash/model/dao/impl/KassaDaoImpl.java
@@ -220,13 +220,13 @@ public class KassaDaoImpl extends AbstractDao implements KassaDao {
      * @throws DaoException ошибке доступа к базе данных или других ошибках.
      */
     @Override
-    public Long countRows() throws DaoException {
-        Long count;
+    public int countRows() throws DaoException {
+        int count;
         try {
             try (PreparedStatement preparedStatement = connection.prepareStatement(SQL_SELECT_COUNT_KASSAS)) {
                 ResultSet resultSet = preparedStatement.executeQuery();
                 resultSet.next();
-                count = resultSet.getLong(1);
+                count = resultSet.getInt(1);
             }
         } catch (SQLException e) {
             logger.log(Level.ERROR, "Database exception during fiend count kassaList row", e);
